v0.3.2
-Refactorisation finished!
It now looks very similar to the class diagram apart from a few
differences (these will be shown in a new updated class diagram
tomorrow)
-New Tokeniser added on.
-Still needs commenting however should be able to do that tomorrow.

Bugs:
WorldGen is giving 7*8 hives, not 7*7. This is causing the world class
to crash when it is trying to tokenise the character array into Cells.